[VT2-39] move: Alarm Service Layer 디렉토리 구조 변경
